Вынес в расширение алерт
